rpi/requirements.txt
requests
Pillow
inky
RPi.GPIO
